Fixed screenshot fields
Fixed the screenshot area of the issue template, as it was set to render as markdown, which made it impossible to paste a screenshot.

Also changed the value from placeholder to value, which will prefill the "zone" with screenshot code

Removed inactive users from the assign attributes of the issue template.