Update add-domain
Add Purolator Canada based phishing domain. I received a text stating I had a package to pickup, despite not having an order with Purolator. The phishing link in full is the bad domain/?UNIQUE_ID . I tested accessing it with Tor but was not able to load any content. This link was received 2023 and is still active, Cloudflare domain registration.