Update falsepositive.list
We build tooling for email marketers, so it is common for our domains to be copy and pasted into Phishing emails in an attempt to make them look legitimate.  We believe this domain was added to your list for that reason.

All content on this domain is generated by paying customers.  Using our service for nefarious purposes is against our terms of service and we would take any report of malicious usage very seriously.  We welcome any reports of malicious content at security@litmus.com

Signed-off-by: Brian Feucht <brian@litmus.com>